+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - FIX RxState sometimes does writes to the storage that do not match the schema.
 - FIX Query Builder does not work on queries created by `.findByIds()` [#6148](https://github.com/pubkey/rxdb/issues/6148)
 - FIX (IndexedDB RxStorage) When used with Sharding and Workers, it sometimes errors on the first query after db creation.
-- 
+
+
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->